enum Type -> dto module, CategoryDTO: String type -> Type type
