+++ dto/src/main/java/com/override/dto/constants/Type.java
@@ -1,0 +1,16 @@
+package com.override.dto.constants;
+
+import lombok.Getter;
+
+@Getter
+public enum Type {
+    INCOME("Доходы"),
+    EXPENSE("Расходы");
+
+    private final String value;
+
+    Type(String value) {
+        this.value = value;
+    }
+
+}
