+++ orchestrator_service/src/main/java/com/override/orchestrator_service/config/DefaultCategoryProperties.java
@@ -1,6 +1,6 @@
 package com.override.orchestrator_service.config;
 
-import com.override.orchestrator_service.constants.Type;
+import com.override.dto.constants.Type;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.boot.context.properties.ConfigurationProperties;
