+++ orchestrator_service/src/main/java/com/override/orchestrator_service/mapper/TransactionMapper.java
@@ -1,7 +1,7 @@
 package com.override.orchestrator_service.mapper;
 
 import com.override.dto.TransactionDTO;
-import com.override.orchestrator_service.constants.Type;
+import com.override.dto.constants.Type;
 import com.override.orchestrator_service.model.Transaction;
 import com.override.dto.TransactionResponseDTO;
 import org.springframework.stereotype.Component;
